REPO-3626: changes after review.
