+++ src/main/resources/alfresco/repository.properties
@@ -623,7 +623,7 @@ content.metadataExtracter.pdf.maxConcurrentExtractionsCount=5
 content.metadataExtracter.pdf.overwritePolicy=PRAGMATIC
 
 #True if bookmarks content should be extracted for PDFBox
-trasformer.pdf.extractBookmarksText=true
+content.transformer.PdfBox.extractBookmarksText=true
 
 # Property to enable upgrade from 2.1-A
 V2.1-A.fixes.to.schema=0
